nptl: hacking to support EDEADLK from kernel
This is not intended to be released. It's probably for the best that
we assert here because I suspect there is a lot of code out there
that wouldn't check the return value and would assume if we return
that we got the mutex, which I don't think we did if the kernel
returned EDEADLK. However, for the specific test I'm doing this is
useful to be able to bubble up to callers.